+++ review-service/src/main/java/com/palja/review_service/exception/ReviewErrorCode.java
@@ -1,0 +1,21 @@
+package com.palja.review_service.exception;
+
+import com.palja.common.exception.ErrorCode;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
+public enum ReviewErrorCode implements ErrorCode {
+
+    NOT_FOUND_REVIEW(HttpStatus.NOT_FOUND, "리뷰를 찾을 수 없습니다");
+
+    private final HttpStatus status;
+    private final String message;
+
+    @Override
+    public HttpStatus getHttpStatus() {
+        return status;
+    }
+}
