--- review-service/src/main/java/com/palja/review_service/infrastructure/repository/ReviewRepositoryAdapter.java
@@ -1,18 +1,0 @@
-package com.palja.review_service.infrastructure.repository;
-
-import com.palja.review_service.domain.entity.Review;
-import com.palja.review_service.domain.repository.ReviewRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class ReviewRepositoryAdapter implements ReviewRepository {
-
-    private final JpaReviewRepository jpaReviewRepository;
-
-    @Override
-    public Review save(Review review) {
-        return jpaReviewRepository.save(review);
-    }
-}
